+++ src/app/fun_batch_d5_016.cpp
@@ -114,4 +114,4 @@ namespace app {
 // ---- sv_animcmd lua_bind wrapper -----------------------------------------
 
 // 0x7101fc11c0
-u64 FUN_7101fc11c0(lua_State *param_1) { app::sv_animcmd::CHECK_FINISH_CAMERA_IF_NOT_HP_MODE(param_1); return 0; }
+u32 FUN_7101fc11c0(lua_State *param_1) { app::sv_animcmd::CHECK_FINISH_CAMERA_IF_NOT_HP_MODE(param_1); return 0; }
